paperpilot-backend-test/pyproject.toml
[tool.poetry]
name = "paperpilot-backend-test"
version = "0.1.0"
description = "paperpilot-backend"
authors = ["Nagico <coyangjr@whu.edu.cn>"]
readme = "README.md"

[[tool.poetry.source]]
name = "aliyun"
priority = "default"
url = "https://mirrors.aliyun.com/pypi/simple/"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
django = "^4.2.1"
django-simpleui = "^2023.8.28"
mysqlclient = "^2.2.0"
pycryptodome = "^3.19.0"
django-split-settings = "^1.2.0"
python-decouple = "^3.8"
loguru = "^0.7.2"
redis = "^5.0.1"
django-extensions = "^3.2.3"
django-cache-url = "^3.4.4"
dj-database-url = "^2.1.0"
hiredis = "^2.2.3"
django-cleanup = "^8.0.0"
django-cacheops = "^7.0.1"
sentry-sdk = "^1.31.0"
pyjwt = "^2.8.0"
alibabacloud-dysmsapi20170525 = "^2.0.24"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.2"
pytest-cov = "^4.1.0"
pytest-django = "^4.5.2"
pytest-mock = "^3.11.1"
model-bakery = "^1.15.0"
pytest-xdist = "^3.3.1"
black = "^23.9.1"
isort = "^5.12.0"
flake8-pyproject = "^1.2.3"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.4.0"
django-debug-toolbar = "^4.2.0"
django-querycount = "^0.8.3"
nplusone = "^1.0.0"
ipython = "^8.16.0"
django-migration-linter = "^5.0.0"
django-extra-checks = "^0.13.3"
pytest = "^7.4.2"
pytest-cov = "^4.1.0"
pytest-django = "^4.5.2"
pytest-mock = "^3.11.1"
model-bakery = "^1.15.0"
pytest-xdist = "^3.3.1"
black = "^23.9.1"
isort = "^5.12.0"
flake8-pyproject = "^1.2.3"
watchfiles = "^0.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 80
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
profile = "black"
skip = ["migrations"]
include_trailing_comma = true
use_parentheses = true
multi_line_output = 3
line_length = 80

[tool.flake8]
ignore = ['E203', 'E266', 'E501', 'W503']
max-line-length = 80
max-complexity = 18
select = ['B', 'C', 'E', 'F', 'W', 'T4']
